Added IPC Api method for external plugins

+++ src/libs/configurator/delete.ts
@@ -1,13 +1,17 @@
 /**
  * @author Chad Koslovsky <chad@technomnancy.it>
  * @file Description
- * @desc Created on 2021-06-29 6:34:49 pm
+ * @desc Created on 2021-06-29 5:57:40 pm
  * @copyright TechnomancyIT
  */
 import Setting from '../../models/Setting';
+import isRenderer from 'is-electron-renderer';
+import { ipcRenderer } from 'electron';
 
 export async function deleteConfig(name: string, values: {}) {
-  return await Setting.destroy({ where: { name, values } }).catch((e) => {
-    error: e;
-  });
+  if (!isRenderer)
+    return await Setting.destroy({ where: { name, values } }).catch((e) => {
+      error: e;
+    });
+  return ipcRenderer.invoke('configurator-delete', { name, values });
 }
